Rename mlkit-related names to kit4dl
