+++ .github/workflows/pip_publish.yml
@@ -1,4 +1,4 @@
-name: Publish MLKit package to PyPI
+name: Publish Kit4DL package to PyPI
 
 on: 
   release:
@@ -7,7 +7,7 @@ on:
 
 jobs:
   build-and-publish:
-    name: Build MLKit and publish to PyPI
+    name: Build Kit4DL and publish to PyPI
     strategy:
       matrix:
         python-version: ["3.10"]
@@ -23,7 +23,7 @@ jobs:
         run: python3 -m pip install build --user
       - name: Build wheels
         run: python -m build --sdist --wheel --outdir=dist/ .
-      - name: Publish MLKit to PyPI
+      - name: Publish Kit4DL to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           password: ${{ secrets.PYPI_API_TOKEN }}
