+++ conda/meta.yaml
@@ -1,43 +1,48 @@
-{% set data = load_file_data('../pyproject.toml', 'toml', from_recipe_dir=True) %}
-{% set prj = data.get('project', {}) %}
+{% set name = "kit4dl" %}
+{% set version = "2023.08a0" %}
 
 package:
-  name: {{ prj.get('name') }}
-  version: {{ prj.get('version') }}
-about:
-  summary: {{ prj.get('description') }}
+  name: "{{ name|lower }}"
+  version: "{{ version }}"
+
 source:
-  path: ...
-  # git_url: https://github.com/opengeokube/ml-kit/
+  path: "https://files.pythonhosted.org/packages/2c/11/1d3a27fb01ee2b182501767a4db60ffa29c67b75908236b84a35d4c37e7b/{{ name }}-{{ version }}.tar.gz"
+  sha256: 07e461adf94f02462b7b3228ea8ded08c6be7a05bc78041e9c89620ae0d59a91
 
 build:
   noarch: python
   number: 0
-  script: python -m pip install --no-deps --ignore-installed ../..
+  script: {{ PYTHON }} -m pip install ../..
 
 requirements:
   build:
-    - python {{ python }}
+    - python >=3.10
   run:
+    - "python >=3.10"
     - "pytorch"
     - "lightning"
     - "torchmetrics"
-    - "numpy>=1.23.0"
-    - "pydantic>=1.10.0"
-    - "scikit-learn>=1.2.0"
+    - "numpy >=1.23.0"
+    - "pydantic >=1.10.0"
+    - "scikit-learn >=1.2.0"
     - "colormath"
     - "tqdm"
     - "typer[all]"
 
-channels:
-  - pytorch 
-  - nvidia 
-  - conda-forge
+test:
+  imports:
+    - kit4dl
+    - kit4dl.tests    
 
 about:
-  home:  https://github.com/opengeokube/ml-kit/
+  home:  https://github.com/opengeokube/kit4dl/
   license: MIT
+  license_file: LICENSE
   summary: 'A quick way to start with machine and deep learning'
-  dev_url: https://github.com/opengeokube/ml-kit/
-  doc_url: https://github.com/opengeokube/ml-kit/blob/main/README.md
-  doc_source_url: https://github.com/opengeokube/ml-kit/blob/main/README.md
+  dev_url: https://github.com/opengeokube/kit4dl/
+  doc_url: https://pypi.org/project/kit4dl
+  doc_source_url: https://github.com/opengeokube/kit4dl/blob/main/README.md
+
+extra:
+  recipe-maintainers:
+      - jamesWalczak
