+++ dev-env.yaml
@@ -1,4 +1,4 @@
-name: mlkit-dev
+name: kit4dl-dev
 channels:
   - pytorch 
   - nvidia
