+++ env.yaml
@@ -1,4 +1,4 @@
-name: mlkit
+name: kit4dl
 channels:
   - pytorch 
   - nvidia
