+++ examples/cnn_s3dis_segmentation/README.md
@@ -13,11 +13,11 @@ Follow the instructions in (http://buildingparser.stanford.edu/dataset.html)[htt
 
 ### 3. Run
 ```bash
-mlkit train
+kit4dl train
 ```
 
 while being in the directory with `conf.toml` file, or specify configuration file explicitly:
 
 ```bash
-mlkit train --conf=<path_to_conf_file>
+kit4dl train --conf=<path_to_conf_file>
 ```
