+++ kit4dl/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Kit4DL CLI sub-package."""
