+++ kit4dl/dataset.py
@@ -1,4 +1,4 @@
-"""A module with the MLKit abstract dataset definition."""
+"""A module with the Kit4DL abstract dataset definition."""
 import logging
 from abc import ABC
 from typing import Any, Callable
@@ -6,9 +6,9 @@ from typing import Any, Callable
 import lightning.pytorch as pl
 from torch.utils.data import DataLoader, Dataset
 
-from mlkit import context
-from mlkit.mixins import LoggerMixin
-from mlkit.nn.confmodels import DatasetConf
+from kit4dl import context
+from kit4dl.mixins import LoggerMixin
+from kit4dl.nn.confmodels import DatasetConf
 
 
 class MLKitAbstractDataModule(ABC, pl.LightningDataModule, LoggerMixin):
@@ -46,7 +46,7 @@ class MLKitAbstractDataModule(ABC, pl.LightningDataModule, LoggerMixin):
             setattr(self, extra_arg_key, extra_arg_value)
 
     def _configure_logger(self) -> None:
-        self._logger = logging.getLogger("mlkit.dataset")
+        self._logger = logging.getLogger("kit4dl.dataset")
         self._logger.setLevel(context.LOG_LEVEL)
         if context.LOG_FORMAT:
             formatter = logging.Formatter(context.LOG_FORMAT)
