+++ kit4dl/mlkit_types.py
@@ -1,4 +1,4 @@
-"""A module with MLKit custom types."""
+"""A module with Kit4DL custom types."""
 from typing import NewType
 
 FullyQualifiedName = NewType("FullyQualifiedName", str)
