+++ kit4dl/nn/__init__.py
@@ -1,0 +1,1 @@
+"""Kit4DL nn subpackage."""
