--- mlkit/__init__.py
@@ -1,3 +1,0 @@
-"""MLKit package."""
-from mlkit.dataset import MLKitAbstractDataModule
-from mlkit.nn.base import MLKitAbstractModule
