--- mlkit/cli/__init__.py
@@ -1,1 +1,0 @@
-"""MLKit CLI sub-package."""
