+++ tests/nn/test_trainer.py
@@ -2,12 +2,12 @@ from unittest.mock import MagicMixin, patch
 
 import pytest
 
-from mlkit.nn.trainer import Trainer, set_seed
+from kit4dl.nn.trainer import Trainer, set_seed
 from tests.fixtures import conf, true_conf
 
 
 class TestTrainer:
-    @patch("mlkit.nn.trainer.set_seed")
+    @patch("kit4dl.nn.trainer.set_seed")
     def test_seed_set_run(self, mock_set_seed, conf):
         Trainer(conf)
         mock_set_seed.assert_called_with(conf.base.seed)
@@ -26,16 +26,16 @@ class TestTrainer:
         assert np.allclose(arr1, arr2)
         assert torch.allclose(ten1, ten2)
 
-    @patch("mlkit.nn.trainer.Trainer._configure_trainer")
-    @patch("mlkit.nn.trainer.Trainer._configure_datamodule")
+    @patch("kit4dl.nn.trainer.Trainer._configure_trainer")
+    @patch("kit4dl.nn.trainer.Trainer._configure_datamodule")
     def test_configure_model(
         self, mock_conf_trainer, mock_conf_datamodule, conf
     ):
         Trainer(conf).prepare()
         conf.model.model_class.assert_called_once_with(conf=conf)
 
-    @patch("mlkit.nn.trainer.Trainer._configure_trainer")
-    @patch("mlkit.nn.trainer.Trainer._configure_model")
+    @patch("kit4dl.nn.trainer.Trainer._configure_trainer")
+    @patch("kit4dl.nn.trainer.Trainer._configure_model")
     def test_configure_datamodule(
         self, mock_conf_trainer, mock_conf_model, conf
     ):
