+++ tests/test_mixins.py
@@ -6,7 +6,7 @@ class TestLoggingMixin:
         "log_method", ["debug", "info", "warn", "error", "critical"]
     )
     def test_log_methods_are_available(self, log_method):
-        from mlkit.mixins import LoggerMixin
+        from kit4dl.mixins import LoggerMixin
 
         class A(LoggerMixin):
             pass
